requirements.txt
fastapi
uvicorn
xgboost
pandas
scikit-learn
faker
requests
python-dotenv
sqlalchemy
